created LogConfigTest class for testing LogConfig specifics
