+++ tests/Integration/Log/LogConfigTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Log;
+
+use Tempest\Log\LogConfig;
+use Tests\Tempest\Integration\FrameworkIntegrationTestCase;
+
+class LogConfigTest extends FrameworkIntegrationTestCase
+{
+    public function test_log_path_by_env(): void
+    {
+        $expectedDebugLogPath = 'log/debug.test.log';
+        $expectedServerLogPath = 'log/server.test.log';
+
+        $this->kernel->loadConfig();
+
+        $logConfig = $this->container->get(LogConfig::class);
+
+        $this->assertSame($expectedDebugLogPath, $logConfig->debugLogPath);
+        $this->assertSame($expectedServerLogPath, $logConfig->serverLogPath);
+    }
+}
